+++ proxy/csv/index.js
@@ -1,5 +1,4 @@
 'use strict';
-const qs = require('querystring');
 const co = require('co');
 const request = require('superagent');
 const csv = require('json2csv');
@@ -8,6 +7,7 @@ const _ = require('lodash');
 const getQueryString = require('helpers/getQueryString.js');
 const csvElements = require('./elements');
 const drivers = require('drivers');
+const listImageRecursive = require('../listImageRecursive');
 const extraRequests = {
   project: (session) => {
     return drivers.keystone.project.listProjectsAsync(
@@ -44,17 +44,6 @@ const extraRequests = {
     );
   }
 };
-function listImageRecursive(query, marker, token, remote, images) {
-  return co(function *() {
-    query = query || {};
-    marker ? query.marker = marker : delete query.marker;
-    let result = (yield drivers.glance.image.listImagesAsync(token, remote, query)).body;
-    images.push.apply(images, result.images);
-    if (result.next) {
-      yield listImageRecursive(query, qs.parse(result.next.split('?')[1]).marker, token, remote, images);
-    }
-  });
-}
 
 module.exports.fields = (req, res, next) => {
   let fields, path = req.path.slice(17);
